rename ShowOtherMergedResult to TransparentMergedResult
